requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
Pillow==10.1.0
python-decouple==3.8
django-filter==23.3
djoser==2.2.0
djangorestframework-simplejwt==5.3.0
